+++ tests/plugins/test_setuptools_pep621.py
@@ -261,8 +261,9 @@ license-files = LICENSE.txt
 
 expected_handle_license_files = """\
 [metadata]
-[metadata.license]
-file = "LICENSE.txt"
+
+[options]
+license-files = ["LICENSE.txt"]
 """
 
 
@@ -271,7 +272,8 @@ def test_handle_license_files(plugin, parse, convert):
     doc = plugin.apply_value_processing(doc)
     print(doc)
     print("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@")
-    doc = plugin.handle_license_and_files(doc)
+    doc = plugin.handle_license(doc)
+    doc = plugin.remove_metadata_not_in_pep621(doc)
     print("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@")
     print(doc)
     assert convert(doc).strip() == expected_handle_license_files.strip()
@@ -284,9 +286,8 @@ license-files = LICENSE.txt, NOTICE.txt
 
 expected_handle_multiple_license_files = """\
 [metadata]
-dynamic = ["license"]
 
-["options.dynamic"]
+[options]
 license-files = ["LICENSE.txt", "NOTICE.txt"]
 """
 
@@ -296,7 +297,8 @@ def test_handle_multiple_license_files(plugin, parse, convert):
     doc = plugin.apply_value_processing(doc)
     print(doc)
     print("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@")
-    doc = plugin.handle_license_and_files(doc)
+    doc = plugin.handle_license(doc)
+    doc = plugin.remove_metadata_not_in_pep621(doc)
     print("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@")
     print(doc)
     assert convert(doc).strip() == expected_handle_multiple_license_files.strip()
@@ -309,11 +311,7 @@ license = MPL-2.0
 
 expected_handle_license = """\
 [metadata]
-dynamic = ["license"]
-
-["options.dynamic"]
-license = "MPL-2.0"
-license-files = ["LICEN[CS]E*", "COPYING*", "NOTICE*", "AUTHORS*"]
+license = {text = "MPL-2.0"}
 """
 
 
@@ -322,7 +320,8 @@ def test_handle_license(plugin, parse, convert):
     doc = plugin.apply_value_processing(doc)
     print(doc)
     print("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@")
-    doc = plugin.handle_license_and_files(doc)
+    doc = plugin.handle_license(doc)
+    doc = plugin.remove_metadata_not_in_pep621(doc)
     print("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@")
     print(doc)
     assert convert(doc).strip() == expected_handle_license.strip()
@@ -330,16 +329,15 @@ def test_handle_license(plugin, parse, convert):
 
 example_handle_license_and_files = """\
 [metadata]
-license = MPL-2.0
+license = MPL-2.0  # comment
 license-files = LICENSE.txt
 """
 
 expected_handle_license_and_files = """\
 [metadata]
-dynamic = ["license"]
+license = {text = "MPL-2.0"} # comment
 
-["options.dynamic"]
-license = "MPL-2.0"
+[options]
 license-files = ["LICENSE.txt"]
 """
 
@@ -349,7 +347,8 @@ def test_handle_license_and_files(plugin, parse, convert):
     doc = plugin.apply_value_processing(doc)
     print(doc)
     print("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@")
-    doc = plugin.handle_license_and_files(doc)
+    doc = plugin.handle_license(doc)
+    doc = plugin.remove_metadata_not_in_pep621(doc)
     print("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@")
     print(doc)
     assert convert(doc).strip() == expected_handle_license_and_files.strip()
@@ -606,14 +605,11 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [project]
-dynamic = ["license", "version"]
+dynamic = ["version"]
 
 [tool]
 [tool.setuptools]
 include-package-data = false
-
-[tool.setuptools.dynamic]
-license-files = ["LICEN[CS]E*", "COPYING*", "NOTICE*", "AUTHORS*"]
 """
 
 
@@ -647,15 +643,12 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [project]
-dynamic = ["license", "version"]
+dynamic = ["version"]
 
 [tool]
 [tool.setuptools]
 data-files = {a = ["b"]}
 include-package-data = false
-
-[tool.setuptools.dynamic]
-license-files = ["LICEN[CS]E*", "COPYING*", "NOTICE*", "AUTHORS*"]
 """